+++ package.json
@@ -70,7 +70,7 @@
     "size-limit": "^11.1.4",
     "ts-morph": "^24.0.0",
     "tsx": "^4.17.0",
-    "typescript": "^5.5.4",
+    "typescript": "5.7.3",
     "typescript-eslint": "^8.1.0",
     "viem": "^2.21.9",
     "vite-tsconfig-paths": "^5.0.1",
