+++ src/core/AbiParameter.ts
@@ -1,16 +1,24 @@
 import type * as abitype from 'abitype'
 
 import * as Errors from './Errors.js'
-import type { AssertName } from './internal/humanReadable/types/signatures.js'
-import type { Modifier } from './internal/humanReadable/types/signatures.js'
+import * as internal from './internal/abiParameter.js'
+import * as internal_signatures from './internal/humanReadable/signatures.js'
+import * as internal_structs from './internal/humanReadable/structs.js'
 import * as internal_regex from './internal/regex.js'
-import type { IsNarrowable, Join } from './internal/types.js'
+import type {
+  FilterReverse,
+  IsNarrowableIncludingNever,
+  Join,
+  TypeErrorMessage,
+} from './internal/types.js'
 
 /** Root type for ABI parameters. */
 export type AbiParameter = abitype.AbiParameter
 
 /**
+ * Formats {@link ox#AbiParameter.AbiParameter} into **Human Readable ABI Parameter**.
  *
+ * TODO
  */
 export function format<
   const abiParameter extends AbiParameter | abitype.AbiEventParameter,
@@ -64,7 +72,10 @@ export declare namespace format {
               [key in keyof components]: format.ReturnType<
                 {
                   type: components[key]['type']
-                } & (IsNarrowable<components[key]['name'], string> extends true
+                } & (IsNarrowableIncludingNever<
+                  components[key]['name'],
+                  string
+                > extends true
                   ? { name: components[key]['name'] }
                   : unknown) &
                   (components[key] extends {
@@ -76,10 +87,10 @@ export declare namespace format {
             },
             ', '
           >})${array}`
-        } & (IsNarrowable<name, string> extends true
+        } & (IsNarrowableIncludingNever<name, string> extends true
           ? { name: name }
           : unknown) &
-          (IsNarrowable<indexed, boolean> extends true
+          (IsNarrowableIncludingNever<indexed, boolean> extends true
             ? { indexed: indexed }
             : unknown)
       >
@@ -88,8 +99,106 @@ export declare namespace format {
         : ''}${abiParameter['name'] extends infer name extends string
         ? name extends ''
           ? ''
-          : ` ${AssertName<name>}`
+          : ` ${internal_signatures.AssertName<name>}`
         : ''}`
+
+  type ErrorType = Errors.GlobalErrorType
+}
+
+/**
+ * Parses **JSON ABI Parameter** or **Human Readable ABI Parameter** into typed {@link ox#AbiParameter.AbiParameter}.
+ *
+ * TODO
+ */
+export function from<
+  const param extends string | readonly string[] | readonly unknown[],
+>(
+  param: abitype.Narrow<param> &
+    (
+      | (param extends string
+          ? param extends ''
+            ? TypeErrorMessage<'Empty string is not allowed.'>
+            : unknown
+          : never)
+      | (param extends readonly string[]
+          ? param extends readonly [] // empty array
+            ? TypeErrorMessage<'At least one parameter required.'>
+            : string[] extends param
+              ? unknown
+              : unknown // TODO: Validate param string
+          : never)
+    ),
+): from.ReturnType<param>
+
+// eslint-disable-next-line jsdoc/require-jsdoc
+export function from(
+  param: string | readonly string[] | readonly unknown[],
+): AbiParameter {
+  if (typeof param === 'string')
+    return internal.parseAbiParameter(param, {
+      modifiers: internal_signatures.modifiers,
+    })
+
+  const abiParameter = (() => {
+    const structs = internal_structs.parseStructs(param as readonly string[])
+    const length = param.length
+    for (let i = 0; i < length; i++) {
+      const signature = (param as readonly string[])[i]!
+      if (internal_signatures.isStructSignature(signature)) continue
+      return internal.parseAbiParameter(signature, {
+        modifiers: internal_signatures.modifiers,
+        structs,
+      })
+    }
+    return undefined
+  })()
+
+  if (!abiParameter) throw new InvalidAbiParameterError({ param })
+
+  return abiParameter
+}
+
+export declare namespace from {
+  type ReturnType<
+    param extends string | readonly string[] | readonly unknown[],
+  > =
+    | (param extends string
+        ? param extends ''
+          ? never
+          : string extends param
+            ? AbiParameter
+            : internal.ParseAbiParameter<
+                param,
+                { modifier: internal_signatures.Modifier }
+              >
+        : never)
+    | (param extends readonly string[]
+        ? string[] extends param
+          ? AbiParameter // Return generic AbiParameter item since type was no inferrable
+          : internal_structs.ParseStructs<param> extends infer structs
+            ? {
+                [key in keyof param]: param[key] extends string
+                  ? internal_signatures.IsStructSignature<
+                      param[key]
+                    > extends true
+                    ? never
+                    : internal.ParseAbiParameter<
+                        param[key],
+                        {
+                          modifier: internal_signatures.Modifier
+                          structs: structs
+                        }
+                      >
+                  : never
+              } extends infer mapped extends readonly unknown[]
+              ? FilterReverse<mapped, never>[0] extends infer result
+                ? result extends undefined
+                  ? never
+                  : result
+                : never
+              : never
+            : never
+        : never)
 
   type ErrorType = Errors.GlobalErrorType
 }
@@ -100,101 +209,6 @@ export class InvalidAbiParameterError extends Errors.BaseError {
   constructor({ param }: { param: string | object }) {
     super('Failed to parse ABI parameter.', {
       details: `parseAbiParameter(${JSON.stringify(param, null, 2)})`,
-    })
-  }
-}
-
-export class InvalidAbiParametersError extends Errors.BaseError {
-  override readonly name = 'AbiParameter.InvalidAbiParametersError'
-
-  constructor({ params }: { params: string | object }) {
-    super('Failed to parse ABI parameters.', {
-      details: `parseAbiParameters(${JSON.stringify(params, null, 2)})`,
-    })
-  }
-}
-
-export class InvalidParameterError extends Errors.BaseError {
-  override readonly name = 'AbiParameter.InvalidParameterError'
-
-  constructor({ param }: { param: string }) {
-    super('Invalid ABI parameter.', {
-      details: param,
-    })
-  }
-}
-
-export class SolidityProtectedKeywordError extends Errors.BaseError {
-  override readonly name = 'AbiParameter.SolidityProtectedKeywordError'
-
-  constructor({ param, name }: { param: string; name: string }) {
-    super('Invalid ABI parameter.', {
-      details: param,
-      metaMessages: [
-        `"${name}" is a protected Solidity keyword. More info: https://docs.soliditylang.org/en/latest/cheatsheet.html`,
-      ],
-    })
-  }
-}
-
-export class InvalidModifierError extends Errors.BaseError {
-  override readonly name = 'AbiParameter.InvalidModifierError'
-
-  constructor({
-    param,
-    type,
-    modifier,
-  }: {
-    param: string
-    type?: abitype.AbiItemType | 'struct' | undefined
-    modifier: Modifier
-  }) {
-    super('Invalid ABI parameter.', {
-      details: param,
-      metaMessages: [
-        `Modifier "${modifier}" not allowed${
-          type ? ` in "${type}" type` : ''
-        }.`,
-      ],
-    })
-  }
-}
-
-export class InvalidFunctionModifierError extends Errors.BaseError {
-  override readonly name = 'AbiParameter.InvalidFunctionModifierError'
-
-  constructor({
-    param,
-    type,
-    modifier,
-  }: {
-    param: string
-    type?: abitype.AbiItemType | 'struct' | undefined
-    modifier: Modifier
-  }) {
-    super('Invalid ABI parameter.', {
-      details: param,
-      metaMessages: [
-        `Modifier "${modifier}" not allowed${
-          type ? ` in "${type}" type` : ''
-        }.`,
-        `Data location can only be specified for array, struct, or mapping types, but "${modifier}" was given.`,
-      ],
-    })
-  }
-}
-
-export class InvalidAbiTypeParameterError extends Errors.BaseError {
-  override readonly name = 'AbiParameter.InvalidAbiTypeParameterError'
-
-  constructor({
-    abiParameter,
-  }: {
-    abiParameter: AbiParameter & { indexed?: boolean | undefined }
-  }) {
-    super('Invalid ABI parameter.', {
-      details: JSON.stringify(abiParameter, null, 2),
-      metaMessages: ['ABI parameter type is invalid.'],
     })
   }
 }
