+++ src/index.ts
@@ -478,6 +478,8 @@ export * as AbiFunction from './core/AbiFunction.js'
 
 /**
  * @category ABI
+ *
+ * TODO
  */
 export * as AbiParameter from './core/AbiParameter.js'
